refactor(reregistration): extract AJAX handler + submission details renderer (S6 of #141)
Pure code movement — no behavior changes. Two focused
collaborators extracted from ReregistrationAdmin (1132 LOC).

  - ReregistrationSubmissionDetailsRenderer (123 LOC) — owns
    build_submission_details_html, the modal HTML builder for the
    "View submission" AJAX endpoint. Promoted from private to
    public so the new AJAX handler can call it.
  - ReregistrationAjaxHandler (142 LOC) — owns the three wp_ajax_*
    callbacks (ajax_generate_ficha, ajax_view_submission_details,
    ajax_count_members). Composes the details renderer via
    constructor and stores it as a private property. The handler
    carries its own CAPABILITY constant; no facade helpers are
    needed because the bodies only touch WP globals + static
    repository/generator classes.

The facade `ReregistrationAdmin::init()` now instantiates the
AJAX handler and registers the three wp_ajax_* hooks against
`array( $this->ajax_handler, '...' )` instead of `array( $this, ... )`.

Public surface preserved: existing tests in
`tests/Unit/ReregistrationAdminTest.php` (and any external
caller) invoke `$admin->ajax_generate_ficha()` /
`->ajax_count_members()` / `->ajax_view_submission_details()`
directly on the facade. To honor that surface without keeping the
bodies on the facade, added three one-line public delegators plus
a `private get_ajax_handler()` that lazily instantiates the
handler when the test calls a delegator before `init()` has run.
Logic still lives entirely in `ReregistrationAjaxHandler`.

Cleanup: removed a stale duplicate copy of `ajax_count_members()`
that earlier work had left in the facade alongside the moved
version.

Deviation: facade is 1002 LOC (above the issue's <800 target).
The issue spec explicitly named multiple large render_* methods
(render_list, render_form, render_submissions,
render_submission_row, render_audience_options,
render_audience_transfer_list) plus handle_save/handle_delete to
keep on the facade. Hitting <800 would require extracting render
helpers the spec did not authorize. Net drop is ~130 LOC for the
two extractions the spec did call out.

Verification:
  - php -l: clean on all 3 files.
  - grep '$this->' on new files: 2 refs, both resolve internally.
  - vendor/bin/phpunit --filter Reregistration: 299/299 OK.
  - vendor/bin/phpunit: 3870/3870 OK.
  - vendor/bin/phpstan analyse includes/reregistration/: no errors.
